+++ src-gui/ui/src/app/app-routing.module.ts
@@ -139,7 +139,7 @@ const routes: Routes = [{
 		path: 'storesetting',
 		component: SettingsComponent,
 		data:{
-			breadcrumb: "Store Settings"
+			breadcrumb: "Settings"
 		}
 	},
 	{